Revert codemagic.tools.google_play.arguments changes
