+++ src/codemagic/tools/google_play/actions/get_latest_build_number_action.py
@@ -23,6 +23,7 @@ class GetLatestBuildNumberAction(GooglePlayBaseAction, metaclass=ABCMeta):
         """
         Get latest build number from Google Play Developer API matching given constraints
         """
+        # TODO: Restore this functionality
         return None
 
     # @cli.action(
