--- src/codemagic/tools/google_play/arguments/__init__.py
@@ -1,5 +1,0 @@
-from .argument_types import CredentialsArgument
-from .google_play_argument import GooglePlayArgument
-from .latest_build_number_argument import LatestBuildNumberArgument
-from .promote_argument import PromoteArgument
-from .tracks_argument import TracksArgument
